train/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.24.0
pandas>=1.5.0
tqdm>=4.65.0
tensorboard>=2.12.0
matplotlib>=3.7.0
scikit-image>=0.20.0
